requirements.txt
appdirs==1.4.4
appnope==0.1.0
argcomplete==1.11.1
asgiref==3.2.7
attrs==19.3.0
awscli==1.18.79
backcall==0.1.0
bleach==3.1.4
boto3==1.14.2
botocore==1.17.2
certifi==2020.4.5.2
cfn-flip==1.2.3
chardet==3.0.4
click==8.0.1
colorama
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.7
django-bootstrap3==12.1.0
dnspython==2.1.0
docutils==0.15.2
durationpy==0.5
email-validator==1.1.2
entrypoints==0.3
filelock==3.0.12
Flask==2.0.1
Flask-WTF==0.15.1
future==0.18.2
gunicorn==20.1.0
hjson==3.0.1
idna
importlib-metadata==4.5.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.15.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kappa==0.6.0
kiwisolver==1.1.0
MarkupSafe==2.0.1
matplotlib==3.2.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pip-tools==5.2.1
placebo==0.9.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
Pygments==2.6.1
pyparsing==2.4.6
pyrsistent==0.16.0
python-dateutil==2.6.1
python-dotenv==0.17.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
svgwrite==1.4
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
tqdm==4.46.1
traitlets==4.3.3
Tree==0.2.4
troposphere==2.6.1
typing-extensions==3.10.0.0
urllib3==1.25.9
virtualenv==20.0.23
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wsgi-request-logger==0.4.6
WTForms==2.3.3
zipp==3.4.1
